% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/featurize.R
\name{feature_frame}
\alias{feature_frame}
\title{Import a bunch of text files as a single data frame}
\usage{
feature_frame(fs)
}
\arguments{
\item{fs}{filenames}
}
\value{
a data frame with two columns, \code{title} and \code{feature}
}
\description{
With one row per feature.
}

