% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/featurize.R
\name{featurize}
\alias{featurize}
\title{Naively featurize lines of text.}
\usage{
featurize(ll, split_pat = "\\\\W+")
}
\arguments{
\item{ll}{lines of text}

\item{split_pat}{a regular expression, handed on to \code{\link{strsplit}}}
}
\value{
a character vector of features
}
\description{
Naively featurize lines of text.
}

