% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laureate_motivations}
\alias{laureate_motivations}
\title{Nobel literature laureate prize citations.}
\format{A character vector, with elements in the same order as the rows
     of \code{laureates}}
\usage{
laureate_motivations
}
\description{
For convenience, the prize ``motivations'' are stored separately here.
The data source is the same as for \code{laureates}
}
\keyword{datasets}

