% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laureates}
\alias{laureates}
\title{Nobel literature laureates.}
\format{A data frame.}
\usage{
laureates
}
\description{
This basic data about the laureates of the Nobel Prize in Literature
was obtained using the data API of \url{http://nobelprize.org}.
}
\details{
For more information, see \url{http://www.nobelprize.org/nobel_organizations/nobelmedia/nobelprize_org/developer/}. According to the site, the data
are ``free to use.''
}
\keyword{datasets}

