% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mallet.R
\name{read_mallet_instances}
\alias{read_mallet_instances}
\title{Read MALLET instances from disk}
\usage{
read_mallet_instances(filename)
}
\arguments{
\item{filename}{name of file to read MALLET instances from}
}
\value{
reference to an instances list object
}
\description{
A helper function for reading a MALLET instance list back into memory.
}
\details{
If you are only loading instances for modeling, you can skip a step and
pass a filename to \code{model$loadDocuments(...)}.
}
\seealso{
\code{\link{write_mallet_instances}}
}

