% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sheik_ll}
\alias{sheik_ll}
\title{The text of E.M. Hull's novel The Sheik.}
\format{A character vector.}
\usage{
sheik_ll
}
\description{
\emph{The Sheik}, by Edith Maude Hull, was published in London by Nash and Grayson in 1919. It is available in a public-domain plain text from Project Gutenberg at \url{http://www.gutenberg.org/cache/epub/7031/pg7031.txt}.
}
\details{
This is a character vector, with one element per line of the plain text file, created by calling

\code{readLines("http://www.gutenberg.org/cache/epub/7031/pg7031.txt")}
}
\keyword{datasets}

