% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/featurize.R
\name{featurize}
\alias{featurize}
\title{Naively featurize lines of text.}
\usage{
featurize(ll)
}
\arguments{
\item{ll}{lines of text}
}
\value{
a character vector of features
}
\description{
Naively featurize lines of text.
}

