% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mallet.R
\name{write_mallet_instances}
\alias{write_mallet_instances}
\title{Save MALLET instances to disk}
\usage{
write_mallet_instances(instances, filename)
}
\arguments{
\item{instances}{instances generated by \code{\link[mallet]{mallet.import}}}

\item{filename}{name of file to save the instance list to}
}
\description{
A helper function for saving a MALLET instance list.
}
\seealso{
\code{\link{read_mallet_instances}}
}

